move chain info helper
